upgrade gqa kernel from test  to mode select
